+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+-  Make `ShopifyApp.configuration.scope` default to empty list `[]` [1913](https://github.com/Shopify/shopify_app/pull/1913)
 
 22.4.0 (August 22, 2024)
 ----------