+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ----------
+
+- Add support for filters in webhook registration [1923](https://github.com/Shopify/shopify_app/pull/1923)
 -  Make `ShopifyApp.configuration.scope` default to empty list `[]` [1913](https://github.com/Shopify/shopify_app/pull/1913)
 
 22.4.0 (August 22, 2024)