+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ----------
 
+22.5.0 (November 28, 2024)
+----------
 - Add support for filters in webhook registration [1923](https://github.com/Shopify/shopify_app/pull/1923)
 -  Make `ShopifyApp.configuration.scope` default to empty list `[]` [1913](https://github.com/Shopify/shopify_app/pull/1913)